cleaned up the image modal code by seperating it into smaller componenets. TraversButtons compoenent and CloseModal Component.
